requirements.txt
Flask==2.3.3
Flask-Cors==4.0.0
redis==5.0.1
python-dotenv==1.0.0
requests==2.31.0
gunicorn==21.2.0

